requirements.txt
fastapi==0.104.1
flask==2.3.3
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
jinja2==3.1.2
joblib==1.3.2
numpy==1.26.0
pandas==2.1.1
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.1
uvicorn==0.23.2
email-validator